chore(assertions): code refactoring
